+++ slirp/if.cpp
@@ -306,7 +306,7 @@ if_start(void)
 	}
 	
 	/* Encapsulate the packet for sending */
-	if_encap((uint8_t*)ifm->m_data, ifm->m_len);
+	if_encap((uint8_t*)ifm->m_data, (int)ifm->m_len);
 
 	m_free(ifm);
